Fixed cassert on MSVC
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@3572 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/DeprecatedGUI/Menu_Net_Host.cpp
@@ -1157,7 +1157,7 @@ void Menu_Net_HostLobbyFrame(int mouse)
 	int secondsTillGameStart = iStartDedicatedSeconds - Round( (tLX->currentTime - fStartDedicatedSecondsPassed).seconds() );
 	if( bStartDedicated && cServer->getNumPlayers() < iStartDedicatedMinPlayers )
 	{
-		if( tLX->currentTime - fStartDedicatedSecondsPassed > iStartDedicatedServerSpamsSomeInfoTimeout )
+		if( tLX->currentTime - fStartDedicatedSecondsPassed > (float)iStartDedicatedServerSpamsSomeInfoTimeout )
 		{
 			cClient->getNetEngine()->SendText( "Game will start when " +
 					itoa(iStartDedicatedMinPlayers) + " players connect", "");
